Update README.md
Link to https://github.com/Cray/chapel-private/issues/3799 (which is about Identify and Implement new solution for Heroku Commit Emailer).

There's always the risk that we implement / replace with a new solution and forget to update the project page. In that case having the link might help someone be less confused. If we do end up keeping the Heroku-based solution and that links adds a little bit of noise but that seems less bad to me than the alternative.